@@ -5,40 +5,28 @@
   "author": "https://github.com/Caphyon",
   "license": "MIT",
   "scripts": {
-    "dev": "cross-env webpack --watch --mode=development --progress --hide-modules",
-    "build": "cross-env webpack --mode=production --progress --hide-modules",
-    "test": "jest"
+    "dev": "vite",
+    "build": "npm run clean-build && vite build && npm run build:copy",
+    "build:copy": "@powershell copy './src/manifest.json' './dist' && @powershell copy './src/assets/icons' './dist/assets' -Recurse && @powershell copy './src/assets/scripts' './dist/assets' -Recurse",
+    "preview": "vite preview",
+    "clean-build": "@powershell If (Test-Path dist) {Remove-Item –path dist -Recurse -Force}"
   },
   "dependencies": {
-    "babel-polyfill": "^6.26.0",
-    "dotenv-webpack": "^1.7.0",
-    "marked": "^4.0.6",
-    "vue": "^2.6.10",
-    "vue-router": "^3.1.3"
+    "buffer": "^6.0.3",
+    "marked": "^9.1.0",
+    "vue": "^3.3.4",
+    "vue-router": "^4.2.5"
   },
   "devDependencies": {
-    "@vue/test-utils": "^1.0.0-beta.29",
-    "babel-core": "^6.26.3",
-    "babel-jest": "^24.9.0",
-    "babel-loader": "^7.1.4",
-    "babel-preset-env": "^1.7.0",
-    "clean-webpack-plugin": "^3.0.0",
-    "copy-webpack-plugin": "^5.0.4",
-    "cross-env": "^6.0.3",
-    "css-loader": "^3.2.0",
-    "file-loader": "^4.2.0",
-    "html-webpack-plugin": "^3.2.0",
-    "jest": "^24.9.0",
-    "jest-fetch-mock": "^3.0.0",
-    "node-sass": "^6.0.1",
-    "sass-loader": "^6.0.1",
-    "uglify-js": "^3.6.1",
-    "uglifyjs-webpack-plugin": "^2.2.0",
-    "vue-jest": "^3.0.5",
-    "vue-loader": "^15.7.1",
-    "vue-template-compiler": "^2.6.10",
-    "webpack": "^4.41.0",
-    "webpack-archive-plugin": "^3.0.0",
-    "webpack-cli": "^3.3.9"
+    "@tailwindcss/nesting": "^0.0.0-insiders.565cd3e",
+    "@vitejs/plugin-vue": "^4.2.3",
+    "autoprefixer": "^10.4.16",
+    "postcss": "^8.4.30",
+    "postcss-nested-ancestors": "^3.0.0",
+    "prettier": "^3.0.3",
+    "prettier-plugin-tailwindcss": "^0.5.5",
+    "sass": "^1.68.0",
+    "tailwindcss": "^3.3.3",
+    "vite": "^4.4.5"
   }
 }
